Extract module name utilities to module_utils.py
Move module name extraction logic from heatmap_collector to shared
module_utils to enable flamegraph to display module names instead of
full file paths.
